feat: add better error messages with clientresponse.c
