+++ src/include/client.h
@@ -9,6 +9,7 @@
 #include <sys/types.h>
 #include "clientargs.h"
 #include "clientrequest.h"
+#include "clientresponse.h"
 
 void handle_get_ok_status(struct client_request_args arg, uint8_t *buf, uint8_t *combinedlen, uint8_t *numeric_data_array, uint32_t *numeric_response);
